+++ app/src/main/java/com/example/cinemalib/model/repository/Repository.kt
@@ -3,6 +3,6 @@ package com.example.cinemalib.model.repository
 import com.example.cinemalib.model.entities.MovieCard
 
 interface Repository {
-    fun getMovieDatafromServer(): MovieCard
+    fun getMovieDataFromServer(): MovieCard
     fun getMovieFromLocalStorage() : List<MovieCard>
 }
