+++ app/src/main/java/com/example/cinemalib/model/repository/RepositoryImpl.kt
@@ -1,11 +1,12 @@
 package com.example.cinemalib.model.repository
 
+import com.example.cinemalib.model.entities.Movie
 import com.example.cinemalib.model.entities.MovieCard
-import com.example.cinemalib.model.entities.getDefaultMovieList
+
 
 class RepositoryImpl : Repository {
-    override fun getMovieDatafromServer() = MovieCard()
+    override fun getMovieDataFromServer() = MovieCard()
 
-    override fun getMovieFromLocalStorage() = getDefaultMovieList()
+    override fun getMovieFromLocalStorage() = Movie.getDefaultMovieList()
 
 }
